% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesPL-package.R
\docType{package}
\name{bayesPL-package}
\alias{bayesPL-package}
\title{bayesPL your description here}
\description{
bayesPL - A package for Bayesian Paired Comparison analysis with Stan
}
\references{

}
