% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpl.R
\name{bpl}
\alias{bpl}
\title{bpl function}
\usage{
bpl(
  x.matrix,
  min = TRUE,
  prior = rep(1, ncol(x.matrix)),
  nsim = 2000,
  nchains = 8,
  seed = 123,
  ...
)
}
\arguments{
\item{data}{input data frame with column y}
}
\description{
bpl function
}
\examples{

}
